version final home, dashboard, aside y tablero
